Fix the "method chaining" calls on Mobiledoc
Turns out `github.com/jbarone/mobiledoc`'s type `Mobiledoc` is a
struct with value receivers, not pointers, and every `.WithXxx` call
mutates the passed-in copy, and then returns the mutated copy.
Except it's programmed in a weird sloppy way where all these copies of
the `Mobiledoc` struct share the `cards` map, because the constructor
initializes it, and thus not chaining `WithCard` calls works, but
`WithAtom` doesn't.

I hate method chaining, and I especially hate it in Go.
Contort the code formatting to unreadable shapes to make it work.